+++ src/PIL/JpegImagePlugin.py
@@ -42,6 +42,7 @@ import subprocess
 import sys
 import tempfile
 import warnings
+from typing import Any
 
 from . import Image, ImageFile
 from ._binary import i16be as i16
@@ -54,7 +55,7 @@ from .JpegPresets import presets
 # Parser
 
 
-def Skip(self, marker):
+def Skip(self: JpegImageFile, marker: int) -> None:
     n = i16(self.fp.read(2)) - 2
     ImageFile._safe_read(self.fp, n)
 
@@ -193,7 +194,7 @@ def APP(self, marker):
             self.info["dpi"] = 72, 72
 
 
-def COM(self, marker):
+def COM(self: JpegImageFile, marker: int) -> None:
     #
     # Comment marker.  Store these in the APP dictionary.
     n = i16(self.fp.read(2)) - 2
@@ -204,7 +205,7 @@ def COM(self, marker):
     self.applist.append(("COM", s))
 
 
-def SOF(self, marker):
+def SOF(self: JpegImageFile, marker: int) -> None:
     #
     # Start of frame marker.  Defines the size and mode of the
     # image.  JPEG is colour blind, so we use some simple
@@ -252,7 +253,7 @@ def SOF(self, marker):
         self.layer.append((t[0], t[1] // 16, t[1] & 15, t[2]))
 
 
-def DQT(self, marker):
+def DQT(self: JpegImageFile, marker: int) -> None:
     #
     # Define quantization table.  Note that there might be more
     # than one table in each marker.
@@ -495,14 +496,14 @@ class JpegImageFile(ImageFile.ImageFile):
 
         self.tile = []
 
-    def _getexif(self):
+    def _getexif(self) -> dict[str, Any] | None:
         return _getexif(self)
 
     def _getmp(self):
         return _getmp(self)
 
 
-def _getexif(self):
+def _getexif(self) -> dict[str, Any] | None:
     if "exif" not in self.info:
         return None
     return self.getexif()._get_merged_dict()