+++ src/PIL/PaletteFile.py
@@ -48,5 +48,5 @@ class PaletteFile:
 
         self.palette = b"".join(self.palette)
 
-    def getpalette(self):
+    def getpalette(self) -> tuple[bytes, str]:
         return self.palette, self.rawmode